[client-test] fix for TestLastErrorEmbeddedInScanTimeoutStatus
Stop injecting a latency on tserver shutdown in the
ClientTest.TestLastErrorEmbeddedInScanTimeoutStatus test, so the tablet
servers are shutting down faster, making the overall test runtime
shorter (~4 vs ~16 seconds).  Besides, it makes it less likely that
some other process binds to the port where the tserver's webserver is
about to re-bind upon restart.

Change-Id: If959b53616a3c0cc1193d57d181f4b0ff034cc83
Reviewed-on: http://gerrit.cloudera.org:8080/7017
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>