[tests] InvalidTokenDuringSeparateWorkloads test
Introduced new AuthnTokenExpireITest.InvalidTokenDuringSeparateWorkloads
test to cover authn token re-acquiring for both read-only and write-only
workloads.

This is to stay clear from the implementation details of the
TestWorkload class: it uses the same client object for both the read and
the write operations, while the authn token is stored as a shared piece
of data in the client object.

Also, renamed AuthnTokenExpireITest.InvalidTokenDuringWorkload test
into AuthnTokenExpireITest.InvalidTokenDuringMixedWorkload.

Change-Id: I7982dfd8456791e57663fc05664e40ff205c42a2
Reviewed-on: http://gerrit.cloudera.org:8080/7025
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins