[tools] Rename `kudu test` to `kudu perf`
This renames the `kudu test` action to `kudu perf`, for the
following reasons:

- The only tool in `kudu test`, `kudu test loadgen`, is meant
to evaluate insertion performance. It's more "perf" than "test".

- There will likely be more perf tools in the future, and other
tools that might fit in the scope of `kudu test`, like ksck, are
located in the action named by thing which they operate on, e.g.
`kudu cluster ksck` and `kudu fs check`.

- There's a file named tool_action_test.cc for the test
action and a file named tool_action-test.cc that contains tests
for tool actions.

Change-Id: I42c53c71c7f2416d421b90997aec6f6513f5f90a
Reviewed-on: http://gerrit.cloudera.org:8080/7117
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>