Temporary workaround for KUDU-1959 (race when selecting rowsets)
As described in the JIRA, there is a race by which multiple MM threads
can race to pick the same rowsets for compaction. Rather than crash when
hitting this bug, it is safe to simply abort that compaction attempt.
The MM will warn about the compaction failure and try again.

This is a temporary workround for the 1.4 release since the issue was
recently reported in the wild on the user list.

Change-Id: I9db313849176e1bf05636d969fafb1682e6d78de
Reviewed-on: http://gerrit.cloudera.org:8080/7120
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
(cherry picked from commit 8be2a59103da46472062f47f89efa6e1bddd0a5c)
Reviewed-on: http://gerrit.cloudera.org:8080/7122
Reviewed-by: Todd Lipcon <todd@apache.org>