Tweak build-and-test.sh to fix spark1 test failures
Maven seems to be caching old artifacts which is leading to
hard-to-diagnose test failures in some branches. This commit changes the
spark1 test to do a full clean/rebuild instead of selectively removing
the kudu-spark target dir. I've also snuck in a change to add the -B
'batch' flag to maven invocations, which should help with the verbose
download progress log spam, as well as the -Dmaven.javadoc.skip flag in
order to speed up the builds, since javadocs aren't necessary.

Change-Id: I92fb0c05a9da65e1a41eb9017050cd8f957ce3a8
Reviewed-on: http://gerrit.cloudera.org:8080/7129
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins