benchmarks: add set -o pipefail
It seems that our automated benchmarks job was failing to build for the
last several days, but we didn't notice because it didn't use
'pipefail'. Thus the 'make | tee build.log' failure didn't actually
cause the script to exit with a bad status code, and we kept running the
binaries from a previous revision.

Change-Id: I6ab66e85cf39c6513da61e472c85e34da77205fc
Reviewed-on: http://gerrit.cloudera.org:8080/7123
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>