KUDU-2037 fix flake in ts_recovery-itest
Fixed flake in TsRecoveryITest.TestRestartWithOrphanedReplicates
scenario.  The write operation timeout was set to 100ms, and for a
ASAN/TSAN build that was under the reasonable minimum to successfully
complete the majority of write operations.

The issue of bloating the client- and the master-side queue with
GetTableLocations() requests will be addressed separately, with a
new integration test to cover the specific issue (see below).

Prior to KUDU-1034 fix, the client continued to retry the operation to
the same tablet server again and again, not invalidating the entry in
its meta-cache.

After KUDU-1034 fix, the client started marking the tserver as failed
and switching to another one, calling GetTableLocations() after every
failure since that was the only available tablet server. In the test
scenario, the master was not responding fast enough to sustain the rate
of adding new entries into the client- and the master-side queues,
so eventually the client timed out on the GetTableLocations() calls.
As a result, the expected tablet crash hadn't happened because there
were too few write operations trigger the crash of the tablet server.

Having short write timeout is not essential for the test.  Bumping
the write operation timeout from 100 to 1000 ms allows for the majority
of write operations to succeed even in TSAN/ASAN builds and avoid
needless retries on the client side.

Change-Id: I6c5449dc9b47062ea9389b25a1b9d906d9de64d9
Reviewed-on: http://gerrit.cloudera.org:8080/7138
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>