ConsensusMetadata::Create() should not overwrite an existing file
We want to be able to control whether clobbering files is allowed. For
example, when there is an existing ConsensusMetadata file on disk, it is
dangerous for Create() to overwrite that file. The safe procedure to
make changes in that situation is to first Load() the file, make
changes, then Flush().

* Stop allowing ConsensusMetadata::Create() to overwrite files.
* Add unit test ConsensusMetadataTest.TestCreateNoOverwrite for this.
* Fix test BootstrapTest.TestOrphanCommit that violated this rule.

Change-Id: I0bfef1ae86a7d2c162095a76bb184a9c18dc69a7
Reviewed-on: http://gerrit.cloudera.org:8080/7190
Tested-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>