fs_manager: clean tmp files from all special directories
Commit c26e09e regressed tmp file cleanup in that the tablet-meta and cmeta
directories were no longer cleaned. This is because I assumed that the
cleanup done in DataDirManager::Open would include these directories (they
are typically found in the first data root), but I forgot that in the
DataDirManager, a data directory is the data root plus a "/data" suffix.

This patch fixes that by explicitly including all "special" directories in
the FsManager's cleanup routine.

Change-Id: Ia6340ad72a6996869e7a83d0d66c165dfe2173c8
Reviewed-on: http://gerrit.cloudera.org:8080/7144
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>