server: consolidate tablet prepare pools
Using the new serial token functionality available in ThreadPool, we can now
safely consolidate all per-tablet prepare pools into a single server-wide
pool. Each replica allocates a token with which to submit to the prepare
pool, attaching the existing tablet-specific metrics to it.

The prepare pool is configured with no upper bound on the number of running
threads because prepare tasks can block on row/schema locks.

Change-Id: Ic393f739cc9c1b267142b20e04a1aaa5aed97cb0
Reviewed-on: http://gerrit.cloudera.org:8080/7150
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>