mini_cluster: shutdown in destructors
The CHECK in the MiniMaster destructor isn't safe; it can fire during test
setup before the master has been shut down. For example, consider this
(edited) code snippet from MiniCluster::StartSingleMaster():

  gscoped_ptr<MiniMaster> mini_master(new MiniMaster(...));
  RETURN_NOT_OK(mini_master->Start());
  RETURN_NOT_OK(mini_master->master()->WaitForSomething(...));
  mini_masters_.push_back(shared_ptr<MiniMaster>(mini_master.release()));

MiniCluster's destructor guarantees that the masters are shut down before
being destroyed, but if WaitForSomething() fails, the new master won't be
added to mini_masters_ and thus will be destroyed without first being shut
down at the end of StartSingleMaster(). To make this safe, let's remove the
CHECK and shut down the master in the destructor. I rolled this out to
MiniTabletServer too, for consistency.

The CHECK (and timeout in StartSingleMaster()) is the cause of one source of
flakiness in delete_tablet-itest.

Change-Id: I742e190f751535d2a59354c3b35cdaf0a340d4ea
Reviewed-on: http://gerrit.cloudera.org:8080/7151
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>