tidy: enable rule for constant naming
This enables the rule that constants should be named kFooBar and fixes
the cases where we didn't adhere to this. The fixes were started using
clang-tidy --fix, but it wasn't very good at finding all references, so
I did some by-hand tweaking to get it to compile after that.

I left a few aberrations in gutil, etc, but this should be at least a
good cleanup.

Change-Id: I7971659ef3152580d44d6ddfb18be7ebf41052c7
Reviewed-on: http://gerrit.cloudera.org:8080/7158
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins