maintenance_manager: allow runtime disabling via flag
This should fix a flake in alter_table-test that was causing failures
about 0.3% of the time when I looped it (and worse than that in the
normal test environment).

The issue was that this test sets FLAGS_enable_maintenance_manager to
false at runtime, and then manually triggers compaction operations and
asserts on the results. In commit 40aa4c3c271c9df20a1 I "fixed" the flag
to only be checked at startup rather than in each iteration of the
maintenance scheduler.

Several other tests also set this flag to false, so it's possible this
will reduce other flakiness as well.

After the fix I looped alter_table-test 2000 times in ASAN with no
failures. Prior to it, it failed 3/1000.

Change-Id: I2207c9674fc1016253af038fc8fa8a14d2bb45a0
Reviewed-on: http://gerrit.cloudera.org:8080/7164
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>