consensus: Get rid of RaftConsensus::Create()
It is no longer useful to have a factory method for RaftConsensus,
particularly since we need to be able to separate creation from
initialization in a follow-up patch. This is a minor refactor that
removes Create() and simplifies a couple of things.

Also adds an additional lifecycle state to RaftConsensus (kNew) for
validation purposes.

Also remove unused MemTracker instance from RaftConsensus.

Change-Id: Ic28fb8fe64cd62d290cea1de22c4ba9dd1743a4e
Reviewed-on: http://gerrit.cloudera.org:8080/7192
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins