[registration-test] fix flake in TestTabletReports
Prior to the fix, the failure rate attributed to this flake was about
0.3% in 1K run:
  http://dist-test.cloudera.org//job?job_id=aserbin.1497582474.17739

After the fix, no failures in 1K run:
  http://dist-test.cloudera.org//job?job_id=aserbin.1497629325.23427

Both runs were of DEBUG build run with -stress_cpu_threads=8 flag.

The crux of the fix is to use AssertEvetually instead of relying on a
hard-coded delay to capture system catalog's metric.  That's because
the metrics of interest are updated upon processing tablet reports from
tservers which contain tablet consensus status information.  The latter
are sent with a tserver-->master heartbeat when the tserver determines
that it has become the leader for one of its tablets.

In addition to that, this patch includes a clean-up of the test code.

Change-Id: I906465ad220236538175c80972ae055193f9bb45
Reviewed-on: http://gerrit.cloudera.org:8080/7209
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>