open FS layout in presence of disk failure
Currently, if a Kudu server starts up with a failed disk, the server
will crash. There are a number of reasons for this, but the pressing
one is that the path instance files may not be readable, meaning the
directories' UUIDs may not be available.

This patch changes this by introducing an "unhealthy" instance in-memory
for each instance that fails to lock, load, canonicalize, etc. Such
instances are ignored when it comes to checking the integrity of the FS
layout, and are simply marked failed by the directory manager.

Testing is done in data_dirs-test, log_block_manager-test, and
fs_manager-test to ensure failed directories do not impede the managers'
startups.

Change-Id: Ib2a1c824526ed52a6b90ddfbc735cecc4c491118
Reviewed-on: http://gerrit.cloudera.org:8080/7784
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>