KUDU-2066. Add experimental Gradle build support
Adds an experimental Gradle build that results in similar
tasks and artifacts as the existing maven build. See the
readme for usage and common commands.

The build is broken out by module with common configurations
in the root build.gradle file. Additionally the gradle
directory contains “drop in” scripts containing functional
pieces of the build configuration to simplify sharing
configurations and the unclutter the main build files.

Includes support for:
- Shaded dependencies that can be used across modules
- Running unit tests before integration tests
- Reporting available dependency updates
- Protobuf and Avro code generation
- Publishing to remote and local maven repositories
- Cross compiling scala
- Code quality checks (checkstyle, findbugs, pmd)

Change-Id: Ib257cdd019a1f383c886b9238bb47d96576c4421
Reviewed-on: http://gerrit.cloudera.org:8080/7258
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins