KUDU-1932. Run at least one tablet-level test against all block managers
Adding current block managers (file and log) to get extra coverage of
block managers. The file block manager especially has some thin coverage
since the log manager is the default on linux machines, such as
dist-test. Additionally, the file block manager is required on macs, so
it's easy for someone working just on a linux machine to break developers who
are using a mac to build and test.

Dist test results:
http://dist-test.cloudera.org/job?job_id=efan.1497991165.6690

Change-Id: I6bf8c4ef45d907f17366776dace9fecd95120821
Reviewed-on: http://gerrit.cloudera.org:8080/7252
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins