kserver: introduce KuduServer class
The kserver module (with the KuduServer class) is an abstraction that sits
between ServerBase and Master/TabletServer. The idea is for it to
encapsulate all Kudu-specific knowledge that should be shared by Masters and
Tablet Servers without polluting ServerBase, which has no Kudu-specific
knowledge and could be reused by any C++ application.

This patch introduces the new class (just lifecycle methods for now) and
wires it into Master/TabletServer. There's no new functionality.

Change-Id: I41ffd09ce46d9d744aa25343cec77b7a33b88563
Reviewed-on: http://gerrit.cloudera.org:8080/7239
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins