[java] Replace interface-annotations with Apache Yetus annotations
Apache Yetus, like our interface-annotations module, was forked
from Hadoop's audience annotations. It does everything
interface-annotations does, using the same vocabulary, and
possibly more, so let's use it instead.

Projects with dependencies on kudu-client or kudu-spark won't
need to change anything in their poms.

This is a binary incompatible change since a jar built after
this change can't be swapped in for a jar built before it-- this
would also require swapping the interface-annotations jar on the
classpath with the appropriate Yetus jar.

I also manually checked that the generated javadocs for the
client look the same as before.

Change-Id: I860e59f497d96f507d5a2b0ea517f9f8d0794dac
Reviewed-on: http://gerrit.cloudera.org:8080/7235
Reviewed-by: Dan Burkert <danburkert@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins