[java] KUDU-2013 re-acquire authn token if expired
This patch introduces automatic authn token re-acquisition when the
current authn token expires.  The client automatically retries the RPC
that hits the token expiration error (the error to re-try is seen as
FATAL_INVALID_AUTHENTICATION_TOKEN sent by the server during connection
negotiation).

Added a few of tests to exercise the new retry logic for automatic
token re-acquisition in case of master-only operations, a bare minimum
workload scenario, and one special case of a connection to the master
opened with secondary credentials.

Change-Id: I0be620629c9a8345ecd5e5679c80ee76ca4eaa57
Reviewed-on: http://gerrit.cloudera.org:8080/7250
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>