tools: add a tool to extract an arbitrary block to stdout
This allows a block to be dumped to stdout. I found this useful when
trying to see how compressible an index block was. We might also find
this useful for debugging block corruption issues on production
clusters -- it's much easier to extract a corrupted block and copy it
around than to try to copy an entire tablet or filesystem.

Change-Id: I04cf75169b1adb721791ca15b2d5ab5067025f06
Reviewed-on: http://gerrit.cloudera.org:8080/7260
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>