fs: add filesystem type detection to data dirs
This patch moves some detection code out of LogBlockManager and into
DataDirManager. It also adds detection of xfs filesystems.

To verify that I didn't regress the existing ext4 on el6 detection behavior,
I ran log_block_manager-test on an ext4 on CentOS 6.6 and verified that the
log output included mention of how the LBM was limiting the containers' max
blocks due to buggy ext4.

Change-Id: I2df259d2fbbc1e950b33005c5da02a3e2cd076f2
Reviewed-on: http://gerrit.cloudera.org:8080/7278
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins