[java] Ensure the shaded packages are always in the same location
Currently when a module depends on kudu-clients and is
shaded it is possible to get bloat in the jar due to
dependencies like guava being included in the shaded
jar twice.

This moves the shade plugin and relocation configuration
to the parent pom's pluginManagement section to ensure the
same packages are always placed in the same location.

For example this reduces the kudu-client-tools jar
from 13.4 MB to 8.5 MB.

Note: The relocation target is changed from
org.apache.kudu.clients.shaded to
org.apache.kudu.shaded since the shading is not specific to
the clients anymore.

Change-Id: I3b519ae75a62ed9bdbf37b7132a43392dd9082db
Reviewed-on: http://gerrit.cloudera.org:8080/7305
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>