consensus_peers: schedule raft heartbeats through messenger
This patch switches Peer to a Messenger-based approach (i.e.
Messenger::ScheduleOnReactor) for scheduling heartbeats. The problem with
ResettableHeartbeater was that it creates a dedicated thread per replica,
and I couldn't see an easy way to share it across replicas.

The scheduling is inspired by libev's "Let the timer time out, but then
re-arm it as required" approach[1]. Callbacks are never canceled since the
Messenger doesn't support that. Instead, the next heartbeat time is
maintained as out-of-band state, and we ensure that the callback period is
low enough that it can honor any heartbeat time.

1. http://pod.tst.eu/http://cvs.schmorp.de/libev/ev.pod#code_ev_timer_code_relative_and_opti

Change-Id: Iac8e09fe02dd32885ef0cf644cb093b1c8e6afb8
Reviewed-on: http://gerrit.cloudera.org:8080/7331
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>