raft_consensus: failure detector doesn't need external synchronization
While working with TimerFailureDetector I saw that it's thread-safe
through and through. Thus, RaftConsensus doesn't need to externally
synchronize it. So I renamed some functions and removed a few lock
acquisition sites.

AFAICT, the only reason not to do this is for one of the overloads of
SnoozeFailureDetector() which calls LOG_WITH_PREFIX_UNLOCKED(). I don't
think that's useful enough to warrant the locking, though.

Change-Id: I637030c3ae2470aa0a8e1067ac6a9b69ea1793b4
Reviewed-on: http://gerrit.cloudera.org:8080/7329
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>