stopwatch: ensure LOG_TIMING always prints
On rare occasion in benchmark runs, the LOG_TIMING macro doesn't produce
any output. It's not entirely clear why this happens, but it seems the
only possibility is that the so-called monotonic clock is going
backwards. This change ensures that the macro always prints even if the
clock moves backwards.

Change-Id: I67fe658801be153cbcb6efb31cc8f3bf9eaf944f
Reviewed-on: http://gerrit.cloudera.org:8080/7333
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>