[java] add BaseKuduTest.doSetup() method
Added BaseKuduTest.doSetup() to call from custom @BeforeClass methods.
This is to allow for run-time customization of MiniKuduCluster
parameters such as number of masters, number of tablet servers and
default timeout for various operations.  The new method is needed for
a follow-up commit.

Also did minor clean-up on some of the related files affected by the
change above.

Change-Id: Id283bcb6a6d109839d86932cb807a1da8156f6eb
Reviewed-on: http://gerrit.cloudera.org:8080/7336
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>