Integrate Mustache templates to webserver
This commit alters the design of webserver path handlers
to use Mustache templates instead of generating HTML
in server code. Path handlers now accept a json object
as a parameter, and fill out the fields of this object
with information to be displayed on their web page.
This json object is then passed to Mustache, along with
the name of the corresponding template, to be rendered.

Many path handlers still conform to the old design of
generating raw HTML; this commit converts two path
handlers to the new design as an example. Converting
old path handlers, and adding new ones, is a simple
process that consists of defining a callback that
accepts a json object and fills out its fields,
passing the function to RegisterPathHandler(), and
adding a new mustache template to www/.

Change-Id: I8cf8181b2b2904a57c9b24ce73316b92a4f6700f
Reviewed-on: http://gerrit.cloudera.org:8080/7448
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>