Add json builder utility
The EasyJson class serves as a wrapper around rapidjson
Value objects, to simplify their usage. It is intended solely
for building json objects, not writing/parsing.

Simplifies code like this:

  rapidjson::Document d;
  rapidjson::Value v;
  v.SetObject();
  rapidjson::Value list;
  list.SetArray();
  v.AddMember("list", list, d.GetAllocator());
  v["list"].PushBack(rapidjson::Value().SetString("element"), d.GetAllocator());

To this:

  EasyJson ej;
  ej["list"][0] = "element";

This commit is intended to facilitate the transition
of webserver path handlers from raw HTML to mustache
templates and json objects.

Change-Id: Ie195b06c5ba67af3043d2032c56b90c1a85f641a
Reviewed-on: http://gerrit.cloudera.org:8080/7387
Reviewed-by: Dan Burkert <danburkert@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins