[tools] Add a 'kudu tablet relocate' tool
This patch adds a relocate tool that moves a tablet replica from
one tablet server to another. Usage:

kudu tablet change_config relocate_replica <masters> <tablet id> <remove id> <add id>

It works by adding <add id> to the configuration, waiting for it
to tablet copy, then removing <remove id>. In the typical case,
this means the number of replicas will change 3 -> 4 -> 3, and
therefore the number of tolerable faults is 1 while the new
tablet bootstraps. As a result of this extra fragility, the tool
requires the tablet to be in "perfect health" when it runs,
meaning ksck returns no errors for the tablet, and also requires
the same after the copy is complete but before removing a
replica. This probably limits the usefulness of the tool to
rebalancing replicas within a healthy tablet.

Additionally, this makes some minimal changes to ksck to allow
it to print to other output streams besides stdout. The purpose
was to allow the output to be suppressed when running the tool,
since the use of ksck is an implementation detail, and the
output is noisy.

Change-Id: I3b7a7243333ba6e6a3d6fce96b220224d6e38a84
Reviewed-on: http://gerrit.cloudera.org:8080/7444
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>